requirements.txt
antlr4-python3-runtime==4.10
black==24.4.2
cfgv==3.4.0
click==8.1.7
colorlog==6.9.0
distlib==0.3.8
exceptiongroup==1.2.2
filelock==3.15.3
identify==2.5.36
iniconfig==2.0.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
pytest==8.3.2
PyYAML==6.0.1
tomli==2.0.1
typing_extensions==4.12.2
virtualenv==20.26.2
